readline/Makefile: adjustment
The symlink is to a directory so -f is an incorrect test.  It will never work.
It doesn't fail because Makefile sees the target 'readline' existing so
  doesn't run it again anyway.

No need for subshell ( ) this is not TCL.
Double dollar to escape Makefile $().
It crudley selects the highest readline version number based on ls ordering.