REWORK: EFHNConcatLook() modifies HierName in place, need thread-safe form
The modification is to make the 'prefix' visible in the hierachy, by making
a temporary modification to hn_parent.  Which works single-threaded but is
not visable multi-threaded.

This looks a huge issue to allow multiple-thread extraction.
Then there is the STDIO removal and buffering (don't use STDIO directly,
 format to memory buffer, then when at checkpoint, like a whole device
 is done, submit to coordinator, and work on next).
Then there is a workload generator / extraction coordinator
Then there is each worker
Then there is the output collation (consistent sort order)

This patch adds a permanent warning with clear indication on the emitted
line what is needed to fix it.

Options:
 * Make HashLookOnly() allow alternative hash/key lookup mechnisms
    (this has other uses in codebase)
    struct keyinfo keyinfo = {
       .arg0_ptr = &tmpHeirNameTreeOnStack,
       .arg1_int = tmpHeirNameCount,
       .method = {  // This inside .rodata
          .hash = &alt_func_hash, /* (Hash*, const struct keyinfo*) */
          .compare = &alt_func_compare /* (const void*, const void*) */
       }
    };
    HashLookOnly(&HashTable, &keyinfo)

Construct the required HierName list on the stack.  Maybe compute the text
string sizes, add padding, add struct HierName, then alloca().  Then fill
in the data.  The HierName path looks to have 1024 limits in place so this
is expected to be under 1024+256 bytes but scales as necessary.