symbols.map: Symbol versioning fixup
Some platforms require all the symbols named in the *.map to be present
in the output being linked.

Maybe there is an option to make it non-fatal, but it is useful to see
what is not present and what is and where from.

The symbols.map.in has a new EOL added so the 'global:' is on a line on
its own.  The filter_symbol_map.pl might be a bit crude for a more
complex *.map file.

Renamed magic/symbol.map to symbol.map.in