netlist.c: etext move to get_etext() function
I think this is trying to work out if the NLNet* is readonly data from the
.text or .rodata or .data segments.  However on a modern system with
DSO loading this whole concept seems a broken idea.  As the heap could
be theoretically be in between with vast areas of unmapped space between
on modern 64bit machine.

The EMSCRIPTEN and __APPLE__ seems broken.  The address of the variable
on the stack is not the same.  __APPLE__ do provide get_etext() API.
On EMSCRIPTEN we return NULL to indicate unknown but also will skip
this method of producing a string label.